Move measurement container to wave_containers dir
Per offline discussion
