+++ wave_containers/CMakeLists.txt
@@ -1,0 +1,27 @@
+CMAKE_MINIMUM_REQUIRED(VERSION 2.8.3)
+PROJECT(wave_containers)
+
+# DEPENDENCIES
+FIND_PACKAGE(Eigen3 REQUIRED)
+
+# INCLUDES
+INCLUDE_DIRECTORIES(
+    include
+    ${Eigen3_DIRECTORY}
+)
+
+# UNIT TESTS
+SET(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/tests)
+
+ADD_EXECUTABLE(
+    # TARGET
+    ${PROJECT_NAME}_tests
+    # TESTS
+    tests/containers/measurement_test.cpp
+    # TEST RUNNER
+    tests/test_runner.cpp
+)
+
+SET(TEST_DEPS gtest)
+
+TARGET_LINK_LIBRARIES(${PROJECT_NAME}_tests ${TEST_DEPS})
