+++ wave_utils/CMakeLists.txt
@@ -39,7 +39,6 @@ ADD_EXECUTABLE(
     tests/utils/filesystem_test.cpp
     tests/utils/math_test.cpp
     tests/utils/time_test.cpp
-    tests/utils/measurement_test.cpp
     # TEST RUNNER
     tests/test_runner.cpp
 )
